+++ src/main/java/com/timskiproekt/pagepicks/repository/UserBookStatusRepository.java
@@ -1,6 +1,6 @@
 package com.timskiproekt.pagepicks.repository;
 
-import com.timskiproekt.pagepicks.model.UserBookStatus;
+import com.timskiproekt.pagepicks.domain.model.UserBookStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,6 +9,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserBookStatusRepository extends JpaRepository<UserBookStatus, Long> {
-    List<UserBookStatus> findByUserId(Integer userId);
-    Optional<UserBookStatus> findByUserIdAndBookIsbn(Integer userId, String bookISBN);
+    List<UserBookStatus> findByUserId(Long userId);
+    Optional<UserBookStatus> findByUserIdAndBookIsbn(Long userId, String bookISBN);
 }
